src/pyproject.toml
[tool.poetry]
name = "poc_cli_translations"
version = "1.0.0"
description = "Basic Python CLI"
authors = ["John Doe <johndoe@mail.com>"]

[tool.poetry.dependencies]
click = "^8.1.3"
python = ">=3.8,<3.11"
python-i18n = "^0.3.9"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-datadir = "^1.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
poc = "poc_cli_translations.cli:main"